fix: move RuntimeLabel to LLVMUtils
